moved entity classes to correct directory
